Added Compression Files
Added Compression Assignments:
Contains LZ.java & MultiNode.java, which perform compression through the
Lempel-Ziv algorithm.  Also included is Huff.java and BinTreeNode.java,
which calculates the how much large a file created by Huffman coding
would create.  While a Huffman tree is created by this program, it was
not created with the intent of creating a compressed file with the
ability to be decompressed.  I plan to add this feature in the near
future.