Uploaded Encyrption Programs
The following files are used for a mock RSA encryption.
KeyGenerator.java creates key.txt, which contains information for both
the public key and private key, should be run before executing RSA.
RSA.java performs an encryption on a provided file to create
"encrypted".  If "encrypted" was the entered file,  It will be decrypted
back to its original format as "decrypted".

Compression/LZ.java was also updated to improve prompt instructions for
the user.